+++ .github/workflows/unit-tests.yml
@@ -43,6 +43,6 @@ jobs:
         run: python -c "from river import datasets; datasets.CreditCard().download(); datasets.Elec2().download(); datasets.Keystroke().download()"
 
       - name: pytest
-        run: pytest --cov=river_torch -m "not datasets"
+        run: pytest --cov=deep_river -m "not datasets"
       - name: Upload coverage reports to Codecov with GitHub Action
         uses: codecov/codecov-action@v2
