+++ .pre-commit-config.yaml
@@ -1,4 +1,4 @@
-files: river_torch
+files: deep_river
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.2.0
