+++ deep_river/anomaly/scaler.py
