+++ deep_river/classification/__init__.py
@@ -1,0 +1,7 @@
+from deep_river.classification.classifier import Classifier
+from deep_river.classification.rolling_classifier import RollingClassifier
+
+__all__ = [
+    "Classifier",
+    "RollingClassifier",
+]
