+++ deep_river/regression/__init__.py
@@ -1,0 +1,7 @@
+from deep_river.regression.regressor import Regressor
+from deep_river.regression.rolling_regressor import RollingRegressor
+
+__all__ = [
+    "Regressor",
+    "RollingRegressor",
+]
