+++ deep_river/utils/estimator_checks.py
