+++ docs/examples/anomaly/example_probability_weighted_autoencoder.md
@@ -1,9 +1,8 @@
 # Probability weighted Autoencoder
 
-
 ```python
 from river import compose, preprocessing, metrics, datasets
-from river_torch.anomaly import ProbabilityWeightedAutoencoder
+from deep_river.anomaly import ProbabilityWeightedAutoencoder
 from torch import nn, manual_seed
 ```
