+++ docs/index.md
@@ -2,5 +2,5 @@
 hide:
     - navigation
 template: home.html
-title: river-torch
+title: deep-river
 ---
