+++ docs/overrides/home.html
@@ -10,7 +10,7 @@
           <div class="tx-hero">
 
             <div class="tx-hero__content">
-                <h1 class=".focus-in-contract-bck"> river-torch </h1>
+                <h1 class=".focus-in-contract-bck"> deep-river </h1>
                 <h3> A Python package for online deep learning that wraps PyTorch for river.</h3>
               <a href="{{ page.next_page.url | url }}" title="{{ page.next_page.title | striptags }}" class="md-button md-button--primary">
                 Get started
