+++ docs/scripts/gen_ref_pages.py
@@ -6,9 +6,9 @@ import mkdocs_gen_files
 
 nav = mkdocs_gen_files.Nav()
 
-for path in sorted(Path("river_torch").rglob("*.py")):
-    module_path = path.relative_to("river_torch").with_suffix("")
-    doc_path = path.relative_to("river_torch").with_suffix(".md")
+for path in sorted(Path("deep_river").rglob("*.py")):
+    module_path = path.relative_to("deep_river").with_suffix("")
+    doc_path = path.relative_to("deep_river").with_suffix(".md")
     full_doc_path = Path("reference", doc_path)
 
     parts = list(module_path.parts)
@@ -27,7 +27,7 @@ for path in sorted(Path("river_torch").rglob("*.py")):
 
     with mkdocs_gen_files.open(full_doc_path, "w+") as fd:
         identifier = ".".join(parts)
-        print(f"::: river_torch.{identifier}", file=fd)
+        print(f"::: deep_river.{identifier}", file=fd)
 
     mkdocs_gen_files.set_edit_path(full_doc_path, path)
