--- river_torch/classification/__init__.py
@@ -1,7 +1,0 @@
-from river_torch.classification.classifier import Classifier
-from river_torch.classification.rolling_classifier import RollingClassifier
-
-__all__ = [
-    "Classifier",
-    "RollingClassifier",
-]
