--- river_torch/regression/__init__.py
@@ -1,7 +1,0 @@
-from river_torch.regression.regressor import Regressor
-from river_torch.regression.rolling_regressor import RollingRegressor
-
-__all__ = [
-    "Regressor",
-    "RollingRegressor",
-]
